chore(settings): refactorered settings design - structure (#135)
* chore(settings): refactorered settings design - structure

* chore(settings): small css fix

* chore(settings): PR comments
