+++ src/account/components/SendPanel/SendPanel.scss
@@ -1,4 +1,4 @@
-.accountPanel {
+.sendPanel {
   .content {
     padding: 0 24px 32px;
   }
