--- src/settings/components/Saved/Saved.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import Icon from 'shared/components/Icon';
-
-export default function Saved() {
-  return (
-    <span>
-      Saved! <Icon name="check" />
-    </span>
-  );
-}
