--- src/settings/components/Saved/index.js
@@ -1,1 +1,0 @@
-export { default } from './Saved.js';
