+++ src/settings/components/Settings/Settings.scss
@@ -1,19 +1,3 @@
 .settings {
-  h1 {
-    margin-top: 0pt;
-  }
-  .divider {
-    margin-top: 10pt;
-  }
-  .networkDetails {
-    margin-top: 10pt;
-    margin-bottom: 10pt;
-  }
-  select {
-    width: 200px;
-    margin: 5px 5px 5px 0;
-    padding: 0.4em;
-    border-radius:10px;
-    border:1px solid #AAAAAA;
-  }
+  padding: 24px;
 }
