+++ src/shared/components/Icon/index.js
@@ -1,1 +1,1 @@
-export { default } from './Icon.js';
+export { default } from './Icon';
